GCP Approach/requirements.txt
requests
pandas
google-cloud-storage
google-cloud-logging
google-cloud-secretmanager